Refactor: Move `AzimuthProvider` to util package
